+++ packages/html/source/application/index.ts
@@ -1,0 +1,1 @@
+export * from './patchOnRaf'
