+++ packages/html/source/domain/services/VNode/getKey.ts
@@ -1,6 +1,7 @@
 import { ComparableValues } from '@typed/lambda'
-import { HtmlVNode, SvgVNode, VNode } from '../model'
-import { isComment, isText } from './VNode'
+import { HtmlVNode, SvgVNode, VNode } from '../../model'
+import { isComment } from './comment'
+import { isText } from './text'
 
 export function getKey<A extends VNode>(vNode: A): ComparableValues | null {
   if (isComment(vNode) || isText(vNode)) {
