+++ packages/html/source/domain/services/VNode/index.ts
@@ -3,3 +3,4 @@ export * from './html'
 export * from './svg'
 export * from './text'
 export * from './vNodesAreEqual'
+export * from './getKey'
