+++ packages/html/source/domain/services/index.ts
@@ -2,5 +2,4 @@ export * from './VNode'
 export * from './diffAttributes'
 export * from './diffRecordMap'
 export * from './EventHandler'
-export * from './getKey'
 export * from './isBooleanAttribute'
