+++ packages/html/source/index.ts
@@ -1,2 +1,3 @@
 export * from './domain'
 export * from './infrastructure'
+export * from './application'
