+++ packages/html/source/infrastructure/index.ts
@@ -1,4 +1,3 @@
-export * from './addElements'
 export * from './constants'
 export * from './createElement'
 export * from './elementToVNode'
@@ -8,7 +7,6 @@ export * from './PatchFailure'
 export * from './removeElements'
 export * from './updateAriaAttributes'
 export * from './updateAttributes'
-export * from './updateChildren'
 export * from './updateDataList'
 export * from './updateEventHandlers'
 export * from './updateProperties'
